examples/resnet/requirements.txt
numpy==2.1.1
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
